Mudando associações e quitando compras
Mudei o modo que os controladores se associam. Agora todos se associam ao controlador do sistema e se comunicam através dele.

+++ entidade/carteira.py
@@ -1,6 +1,6 @@
 class Carteira:
     def __init__(self, dinheiro):
-        self.__dinheiro = int(dinheiro)
+        self.__dinheiro = float(dinheiro)
 
     @property
     def dinheiro(self):
